pyproject.toml
[tool.poetry]
name = "tradeweb"
version = "0.1.0"
description = ""
authors = ["Alex Georgiou <alex.georgio@ymail.com>"]

[tool.poetry.dependencies]
python = "3.9"
numpy = "1.23.4"
pytz = "2022.6"
PyYAML = "^6.0"
kafka-python = "^2.0.2"
pytest = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
